requirements.txt
elasticsearch
pytz
xlrd
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.9
requests==2.23.0
schedule==0.6.0
soupsieve==2.0
urllib3==1.25.8
link_preview
PyYAML